refactor(step3): abstract step content
