+++ ui/src/components/forms/step3/RoadTripContent.jsx
@@ -1,0 +1,10 @@
+const RoadTripContent = () => {
+
+  return (
+    <div>
+      
+    </div>
+  );
+};
+
+export default RoadTripContent;
