+++ ui/src/components/forms/step3/ZoneTripContent.jsx
@@ -1,0 +1,10 @@
+const ZoneTripContent = () => {
+
+    return (
+      <div>
+        
+      </div>
+    );
+  };
+  
+  export default ZoneTripContent;
